+++ aa.txt
@@ -1,0 +1,31 @@
+/mnt/d/tmp/DQNTrainer_11580_0103.log:2020-01-03 11:34:18,910 INFO	APP[MarketDirClassifier.O10001] loaded e:/AShareSample/ETF/RFrames_SH510300.h5 and found 57 ReplayFrames
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:12:19,738 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.24480408197268844] accu[0.9419403] from eval-result [0.26947679778095335, 0.9314575]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:18:22,130 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.23911389254499227] accu[0.9464264] from eval-result [0.29522238654317334, 0.9236908]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:24:29,308 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.22828572569414973] accu[0.94877625] from eval-result [0.22273056535050273, 0.9508209]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:30:43,318 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.22904157231096178] accu[0.95033264] from eval-result [0.22762960247928277, 0.9496765]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:37:31,021 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.21573409694246948] accu[0.9569855] from eval-result [0.20646846399176866, 0.9595642]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:43:45,581 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.2205458638491109] accu[0.95196533] from eval-result [0.21564055618364364, 0.95254517]
+/mnt/d/tmp/DQNTrainer_13360_0103.log:2020-01-03 10:50:00,906 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_13360\VGG16d1.S1548I4A3.weights.h5, loss[0.23354167107027024] accu[0.94555664] from eval-result [0.2332830864470452, 0.9430084]
+
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:01:02,678 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.21876043872907758] accu[0.95469666] from eval-result [0.32141214434523135, 0.9202728]
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:07:05,342 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.21730503451544791] accu[0.9559326] from eval-result [0.37378665897995234, 0.9190521]
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:13:09,734 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.24665950424969196] accu[0.9398651] from eval-result [0.4238125014817342, 0.87327576]
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:19:11,616 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.22494955186266452] accu[0.95240784] from eval-result [0.42792135500349104, 0.88194275]
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:25:14,744 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.2504401500336826] accu[0.941803] from eval-result [0.45918020652607083, 0.8636322]
+/mnt/d/tmp/DQNTrainer_4268_0103.log:2020-01-03 11:31:21,694 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_4268\VGG16d1.S1548I4A3.weights.h5, loss[0.23875526536721736] accu[0.94577026] from eval-result [0.459727352950722, 0.8663025]
+
+
+
+
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 11:42:04,813 INFO	APP[MarketDirClassifier.O10001] loaded e:/AShareSample/ETF/RFrames_SH510050.h5 and found 57 ReplayFrames
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 11:51:59,793 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.21736193157266825] accu[0.9533844] from eval-result [0.24025319324573502, 0.9448242]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 11:58:00,304 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22164087498094887] accu[0.95266724] from eval-result [0.2855838545365259, 0.9262848]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:04:04,663 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22138105717021972] accu[0.95321655] from eval-result [0.26329224347136915, 0.9363556]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:10:10,630 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.23856768081896007] accu[0.9436188] from eval-result [0.33647539804223925, 0.90444946]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:16:15,746 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22929421719163656] accu[0.9504852] from eval-result [0.34629059419967234, 0.9007263]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:22:17,056 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22652861766982824] accu[0.9505005] from eval-result [0.3701532611157745, 0.89604187]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:28:18,156 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.2347871670499444] accu[0.9480896] from eval-result [0.3876365418545902, 0.8851471]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:34:22,231 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.21488622191827744] accu[0.95700073] from eval-result [0.2521765986457467, 0.9420929]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:40:26,032 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22809107939247042] accu[0.95111084] from eval-result [0.2944069429067895, 0.92256165]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:46:30,713 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.21278859116137028] accu[0.95954895] from eval-result [0.2696174131706357, 0.93655396]
+/mnt/d/tmp/DQNTrainer_9656_0103.log:2020-01-03 12:52:35,699 INFO	APP[MarketDirClassifier.O10001] doAppStep_local_generator() done, saved weights ./out/DQNTrainer_9656\VGG16d1.S1548I4A3.weights.h5, loss[0.22916756267659366] accu[0.950943] from eval-result [0.49251457198988646, 0.84539795]
